LIR: Drop lowering transformation
This should better be done during SMT encoding.

An example where this might break is when during the construction
of the counterexample, expressions from HIR are evaluated. Because
the HIR expressions were unaffected by this transformation, "unencodable"
operators, such as bool2bv, ended up in Solver/RSMT and caused a panic.